Ensure only one row is returned
We recently found that recovering an actor with large sized snapshots failed due to a missing limit on the query. The query as is stands returns every snapshot meeting the criteria which for us was several 10s of GBs of data. There was an existing limit on  _selectByPersistenceIdAndMaxSeqNr, but none on the other 3 similar queries where a max sequence number or timestamp are specified